feat: add shared package for reused pages
