+++ apps/social/next.config.mjs
@@ -8,7 +8,7 @@ const nextConfig = {
     removeConsole: process.env.NODE_ENV === "production",
   },
 
-  transpilePackages: ["@umamin/ui", "@umamin/db", "@umamin/gql"],
+  transpilePackages: ["@umamin/ui", "@umamin/db", "@umamin/gql", "@umamin/shared"],
   images: {
     remotePatterns: [
       {
