+++ apps/social/src/app/(authentication)/login/google/route.ts
@@ -1,31 +1,1 @@
-import { generateState, generateCodeVerifier } from "arctic";
-import { cookies } from "next/headers";
-import { google } from "@/lib/auth";
-
-export async function GET(): Promise<Response> {
-  const state = generateState();
-  const codeVerifier = generateCodeVerifier();
-  const url = await google.createAuthorizationURL(state, codeVerifier, {
-    scopes: ["profile", "email"],
-  });
-
-  url.searchParams.set("access_type", "offline");
-
-  cookies().set("google_oauth_state", state, {
-    path: "/",
-    secure: process.env.NODE_ENV === "production",
-    httpOnly: true,
-    maxAge: 60 * 10,
-    sameSite: "lax",
-  });
-
-  cookies().set("code_verifier", codeVerifier, {
-    path: "/",
-    secure: process.env.NODE_ENV === "production",
-    httpOnly: true,
-    maxAge: 60 * 10,
-    sameSite: "lax",
-  });
-
-  return Response.redirect(url);
-}
+export { GET } from "@umamin/shared/routes/google/route";
