+++ packages/shared/postcss.config.cjs
@@ -1,0 +1,1 @@
+module.exports = require("@umamin/ui/postcss.config");
