+++ packages/shared/tailwind.config.ts
@@ -1,0 +1,1 @@
+export * from "@umamin/ui/tailwind.config";
